Add new command for new controllers
